Improved sshd enable check

+++ build/my_init.d/15_sshd_init.sh
@@ -2,8 +2,10 @@
 
 if [ $SSHD_ENABLED -eq 1 ]; then
   echo "### sshd service enabled"
-  rm -f /etc/service/sshd/down
-  /etc/my_init.d/00_regen_ssh_host_keys.sh
+  if [ -e /etc/service/sshd/down ]; then
+    rm -f /etc/service/sshd/down
+    /etc/my_init.d/00_regen_ssh_host_keys.sh
+  fi
 else
   echo "### sshd service disabled"
   touch /etc/service/sshd/down
